# openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MessageCreate:
      properties:
        content:
          title: Content
          type: string
        is_broadcast:
          default: false
          title: Is Broadcast
          type: boolean
        message_type:
          default: message
          title: Message Type
          type: string
        priority:
          default: normal
          title: Priority
          type: string
        recipient_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Recipient Id
        subject:
          title: Subject
          type: string
      required:
      - subject
      - content
      title: MessageCreate
      type: object
    MessageResponse:
      properties:
        content:
          title: Content
          type: string
        created_at:
          title: Created At
          type: string
        id:
          title: Id
          type: integer
        is_broadcast:
          title: Is Broadcast
          type: boolean
        is_read:
          title: Is Read
          type: boolean
        message_type:
          title: Message Type
          type: string
        priority:
          title: Priority
          type: string
        read_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Read At
        recipient_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Recipient Email
        recipient_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Recipient Id
        sender_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Sender Email
        sender_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sender Id
        subject:
          title: Subject
          type: string
      required:
      - id
      - sender_id
      - recipient_id
      - message_type
      - subject
      - content
      - is_read
      - is_broadcast
      - priority
      - created_at
      - read_at
      - sender_email
      - recipient_email
      title: MessageResponse
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  description: Backend API for cybercrime investigation dashboard
  title: Cybercrime Investigation API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/messages/broadcast:
    post:
      description: Broadcast an announcement to all investigators
      operationId: broadcast_announcement_api_v1_messages_broadcast_post
      parameters:
      - in: query
        name: sender_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sender Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Broadcast Announcement Api V1 Messages Broadcast Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Broadcast Announcement
      tags:
      - messages
  /api/v1/messages/investigators/{investigator_id}/message:
    post:
      description: Send a message to a specific investigator
      operationId: send_message_to_investigator_api_v1_messages_investigators__investigator_id__message_post
      parameters:
      - in: path
        name: investigator_id
        required: true
        schema:
          title: Investigator Id
          type: integer
      - in: query
        name: sender_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sender Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Send Message To Investigator
      tags:
      - messages
  /api/v1/messages/investigators/{investigator_id}/messages:
    get:
      description: Get all messages for a specific investigator
      operationId: get_investigator_messages_api_v1_messages_investigators__investigator_id__messages_get
      parameters:
      - in: path
        name: investigator_id
        required: true
        schema:
          title: Investigator Id
          type: integer
      - in: query
        name: unread_only
        required: false
        schema:
          default: false
          title: Unread Only
          type: boolean
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MessageResponse'
                title: Response Get Investigator Messages Api V1 Messages Investigators  Investigator
                  Id  Messages Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Investigator Messages
      tags:
      - messages
  /api/v1/messages/investigators/{investigator_id}/reply:
    post:
      description: Investigator replies to a message from superadmin
      operationId: reply_to_message_api_v1_messages_investigators__investigator_id__reply_post
      parameters:
      - in: path
        name: investigator_id
        required: true
        schema:
          title: Investigator Id
          type: integer
      - in: query
        name: message_id
        required: true
        schema:
          title: Message Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reply To Message
      tags:
      - messages
  /api/v1/messages/investigators/{investigator_id}/unread-count:
    get:
      description: Get count of unread messages for an investigator
      operationId: get_unread_count_api_v1_messages_investigators__investigator_id__unread_count_get
      parameters:
      - in: path
        name: investigator_id
        required: true
        schema:
          title: Investigator Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Unread Count
      tags:
      - messages
  /api/v1/messages/messages/{message_id}/read:
    patch:
      description: Mark a message as read
      operationId: mark_message_as_read_api_v1_messages_messages__message_id__read_patch
      parameters:
      - in: path
        name: message_id
        required: true
        schema:
          title: Message Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Mark Message As Read
      tags:
      - messages
...